Application V1

+++ app/src/main/java/com/example/flori/geoguessr/Score.java
@@ -5,14 +5,14 @@ import com.orm.SugarRecord;
 import java.util.Date;
 
 public class Score extends SugarRecord<Score> {
-    int score;
+    float score;
     String level;
     Date date;
 
     public Score() {
     }
 
-    public Score(int score, String level, Date date) {
+    public Score(float score, String level, Date date) {
         this.score = score;
         this.level = level;
         this.date = date;
